refactor: move datev-specific stuff to utils
